% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree.R
\name{blackboost_train}
\alias{blackboost_train}
\title{Boosted trees via mboost}
\usage{
blackboost_train(
  formula,
  data,
  family,
  weights = NULL,
  teststat = "quad",
  testtype = "Teststatistic",
  mincriterion = 0,
  minsplit = 10,
  minbucket = 4,
  maxdepth = 2,
  saveinfo = FALSE,
  ...
)
}
\arguments{
\item{...}{Other arguments to pass.}

\item{x}{A data frame or matrix of predictors.}

\item{y}{A factor vector with 2 or more levels}
}
\value{
A fitted blackboost model.
}
\description{
\code{blackboost_train} is a wrapper for the \code{blackboost()} function in the
\pkg{mboost} package that fits tree-based models
where all of the model arguments are in the main function.
}
\keyword{internal}
