% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored-package.R
\docType{package}
\name{censored-package}
\alias{censored}
\alias{censored-package}
\title{censored: Parsnip Engines for Survival Models}
\description{
\code{censored} provides engines for survival models from the \code{parsnip} package.
}
\examples{
library(survival)

# Accelerated Failure Time (AFT) model

fit_aft <- survival_reg(dist = "weibull") \%>\%
  set_engine("survival") \%>\%
  fit(Surv(time, status) ~ age + sex + ph.karno, data = lung)
predict(fit_aft, lung[1:3, ], type = "time")

# Cox's Proportional Hazards model

fit_cox <- proportional_hazards() \%>\%
  set_engine("survival") \%>\%
  fit(Surv(time, status) ~ age + sex + ph.karno, data = lung)
predict(fit_cox, lung[1:3, ], type = "time")

# Andersen-Gill model for recurring events

fit_ag <- proportional_hazards() \%>\%
  set_engine("survival") \%>\%
  fit(Surv(tstart, tstop, status) ~ treat + inherit + age, data = cgd)
predict(fit_ag, cgd[1:3, ], type = "time")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/censored}
  \item Report bugs at \url{https://github.com/tidymodels/censored/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Authors:
\itemize{
  \item Hannah Frick \email{hannah@rstudio.com}
}

}
\keyword{internal}
