% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{cond_inference_surv_ctree}
\alias{cond_inference_surv_ctree}
\title{A wrapper function for conditional inference trees}
\usage{
cond_inference_surv_ctree(
  formula,
  data,
  minsplit = 20,
  maxdepth = 0,
  teststat = "quad",
  testtype = "Bonferroni",
  mincriterion = 0.95,
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit. Note that
symbols like \code{:} and \code{-} will not work and the tree will make use
of all variables listed on the right-hand side of \code{formula}.}

\item{data}{a data frame containing the variables in the model.}

\item{minsplit}{the minimum sum of weights in a node in order to be
considered for splitting.}

\item{maxdepth}{maximum depth of the tree. The default \code{maxdepth = 0}
means that no restrictions are applied to tree sizes.}

\item{teststat}{a character specifying the type of the test statistic to be
applied.}

\item{testtype}{a character specifying how to compute the distribution of
the test statistic.}

\item{mincriterion}{the value of the test statistic (for \code{testtype ==
"Teststatistic"}), or 1 - p-value (for other values of \code{testtype}) that
must be exceeded in order to implement a split.}

\item{...}{Other options to pass to \code{\link[party:ctree]{party::ctree()}}.}
}
\value{
An object of class \code{BinaryTree-class}.
library(survival)
data(lung)
cond_inference_surv_ctree(
Surv(time, status) ~ age + sex,
data = lung,
maxdepth = 2,
minsplit = 5
)
}
\description{
This function is a slightly different API for \code{\link[party:ctree]{party::ctree()}} that has
several important arguments as top-level arguments (as opposed to being
specified in \code{\link[party:ctree_control]{party::ctree_control()}}).
}
\seealso{
cond_inference_surv_cforest
}
