% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_survival_prop.R
\name{coxnet_survival_prob}
\alias{coxnet_survival_prob}
\title{A wrapper for survival probabilities with coxnet models}
\usage{
coxnet_survival_prob(object, new_data, times, output = "surv", ...)
}
\arguments{
\item{object}{A fitted \verb{_coxnet} object.}

\item{new_data}{Data for prediction.}

\item{times}{A vector of integers for prediction times.}

\item{output}{One of "surv" or "haz".}

\item{...}{Options to pass to \code{\link[survival:survfit]{survival::survfit()}}.}
}
\value{
A nested tibble.
}
\description{
A wrapper for survival probabilities with coxnet models
}
\keyword{internal}
