% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_hazards-glmnet.R
\name{coxnet_train}
\alias{coxnet_train}
\title{Wrapper for glmnet for censored}
\usage{
coxnet_train(
  formula,
  data,
  alpha = 1,
  lambda = NULL,
  weights = NULL,
  ...,
  call = caller_env()
)
}
\arguments{
\item{formula}{The model formula.}

\item{data}{The data.}

\item{...}{additional parameters passed to glmnet::glmnet.}

\item{call}{The call passed to \code{\link[rlang:abort]{rlang::abort()}}.}
}
\value{
A fitted \code{glmnet} model.
}
\description{
Not to be used directly by users.
}
\details{
This wrapper translates from formula interface to glmnet's matrix due to how
stratification can be specified. glmnet requires that the \emph{response} is
stratified via \code{\link[glmnet:stratifySurv]{glmnet::stratifySurv()}}. censored allows specification via a
\code{\link[survival:strata]{survival::strata()}} term on the right-hand side of the formula. The formula
is used to generate the stratification information needed for stratifying the
response. The formula without the strata term is used for generating the
model matrix for glmnet.

The wrapper retains the original formula and the pre-processing elements
including the training data to allow for predictions from the fitted model.
}
\examples{
coxnet_mod <- coxnet_train(Surv(time, status) ~ age + sex, data = lung)
}
\keyword{internal}
