% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_survival_prop.R
\name{cph_survival_prob}
\alias{cph_survival_prob}
\title{A wrapper for survival probabilities with cph models}
\usage{
cph_survival_prob(x, new_data, times, output = "surv", conf.int = 0.95, ...)
}
\arguments{
\item{x}{A model from \code{coxph()}.}

\item{new_data}{Data for prediction}

\item{times}{A vector of integers for prediction times.}

\item{output}{One of "surv", "conf", or "haz".}

\item{conf.int}{The confidence level}

\item{...}{Options to pass to \code{\link[survival:survfit]{survival::survfit()}}}
}
\value{
A nested tibble
}
\description{
A wrapper for survival probabilities with cph models
}
\keyword{internal}
