% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_tree_ipred.R
\name{details_bag_tree_ipred}
\alias{details_bag_tree_ipred}
\title{Ensembles of decision trees}
\description{
\code{\link[ipred:bagging]{ipred::bagging()}} fits an ensemble of decision trees.
}
\details{
For this engine, there is a single mode: censored regression
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package (censored regression)}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{bag_tree(engine = "ipred") \%>\% 
  set_mode("censored regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Bagged Decision Tree Model Specification (censored regression)
## 
## Main Arguments:
##   cost_complexity = 0
##   min_n = 2
## 
## Computational engine: ipred 
## 
## Model fit template:
## ipred::bagging(formula = missing_arg(), data = missing_arg())
}
}

\subsection{Preprocessing requirements}{

FIXME
}

\subsection{References}{
\itemize{
\item Breiman L. 1996. “Bagging predictors”. Machine Learning. 24 (2):
123-140
\item Hothorn T, Lausen B, Benner A, Radespiel-Troeger M. 2004. Bagging
Survival Trees. \emph{Statistics in Medicine}, 23(1), 77–91.
\item Kuhn, M, and K Johnson. 2013. \emph{Applied Predictive Modeling}.
Springer.
}
}
}
\keyword{internal}
