% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_hazards_glmnet.R
\name{details_proportional_hazards_glmnet}
\alias{details_proportional_hazards_glmnet}
\title{Proportional hazards regression}
\description{
\code{\link[glmnet:glmnet]{glmnet::glmnet()}} fits a regularized Cox proportional hazards model.
}
\details{
For this engine, there is a single mode: censored regression
\subsection{Tuning Parameters}{

This model has 2 tuning parameters:
\itemize{
\item \code{penalty}: Amount of Regularization (type: double, default: see
below)
\item \code{mixture}: Proportion of Lasso Penalty (type: double, default: 1.0)
}

A value of \code{mixture = 1} corresponds to a pure lasso model, while
\code{mixture = 0} indicates ridge regression.

The \code{penalty} parameter has no default and requires a single numeric
value. For more details about this, and the \code{glmnet} model in general,
see \link[parsnip:glmnet-details]{parsnip::glmnet-details}.
}

\subsection{Translation from parsnip to the original package}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{proportional_hazards(penalty = double(1), mixture = double(1)) \%>\% 
  set_engine("glmnet") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Proportional Hazards Model Specification (censored regression)
## 
## Main Arguments:
##   penalty = 0
##   mixture = double(1)
## 
## Computational engine: glmnet 
## 
## Model fit template:
## censored::glmnet_fit_wrapper(formula = missing_arg(), data = missing_arg(), 
##     family = missing_arg(), alpha = double(1))
}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine. When using the
formula method via
\code{\link[=fit.model_spec]{fit.model_spec()}}, parsnip will
convert factor columns to indicators.

Predictors should have the same scale. One way to achieve this is to
center and scale each so that each predictor has mean zero and a
variance of one. By default, \code{\link[glmnet:glmnet]{glmnet::glmnet()}} uses
the argument \code{standardize = TRUE} to center and scale the data.
}
}
\section{References}{
\itemize{
\item Simon N, Friedman J, Hastie T, Tibshirani R. 2011. “Regularization
Paths for Cox’s Proportional Hazards Model via Coordinate Descent.”
\emph{Journal of Statistical Software}, Articles 39 (5): 1–13. .
\item Hastie T, Tibshirani R, Wainwright M. 2015. \emph{Statistical Learning
with Sparsity}. CRC Press.
\item Kuhn M, Johnson K. 2013. \emph{Applied Predictive Modeling}. Springer.
}
}

\keyword{internal}
