% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_hazards_survival.R
\name{details_proportional_hazards_survival}
\alias{details_proportional_hazards_survival}
\title{Proportional hazards regression}
\description{
\code{\link[survival:coxph]{survival::coxph()}} fits a Cox proportional hazards model.
}
\details{
For this engine, there is a single mode: censored regression
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{proportional_hazards() \%>\% 
  set_engine("survival") \%>\% 
  set_mode("censored regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Proportional Hazards Model Specification (censored regression)
## 
## Computational engine: survival 
## 
## Model fit template:
## survival::coxph(formula = missing_arg(), data = missing_arg(), 
##     x = TRUE, model = TRUE)
}
}

\subsection{References}{
\itemize{
\item Andersen P, Gill R. 1982. Cox’s regression model for counting
processes, a large sample study. \emph{Annals of Statistics} 10,
1100-1120.
}
}
}
\keyword{internal}
