% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_reg_survival.R
\name{details_survival_reg_survival}
\alias{details_survival_reg_survival}
\title{Parametric survival regression}
\description{
\code{\link[survival:survreg]{survival::survreg()}} fits a parametric survival model.
}
\details{
For this engine, there is a single mode: censored regression
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{survival_reg() \%>\% 
  set_engine("survival") \%>\% 
  set_mode("censored regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Parametric Survival Regression Model Specification (censored regression)
## 
## Computational engine: survival 
## 
## Model fit template:
## survival::survreg(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg(), model = TRUE)
}
}

\subsection{Other details}{

Note that \code{model = TRUE} is needed to produce quantile predictions when
there is a stratification variable and can be overridden in other cases.
}

\subsection{References}{
\itemize{
\item Kalbfleisch, J. D. and Prentice, R. L. 2002 \emph{The statistical
analysis of failure time data}, Wiley.
}
}
}
\keyword{internal}
