% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_reg.R
\name{flexsurv_probs}
\alias{flexsurv_probs}
\alias{survreg_survival_probs}
\alias{survreg_hazard_probs}
\title{Internal function helps for parametric survival models}
\usage{
flexsurv_probs(object, new_data, time, type = "survival")

survreg_survival_probs(object, new_data, time)

survreg_hazard_probs(object, new_data, time)
}
\arguments{
\item{object}{A \code{survreg} or \code{flexsurvreg} object.}

\item{new_data}{A data frame.}

\item{time}{A vector of time points}
}
\value{
A nested tibble with column name \code{.pred}
}
\description{
Internal function helps for parametric survival models
}
\keyword{internal}
