% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_reg_data.R
\name{glmnet_fit_wrapper}
\alias{glmnet_fit_wrapper}
\title{Wrapper for glmnet for survnip}
\usage{
glmnet_fit_wrapper(x, y, ...)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation
vector. Can be in sparse matrix format (inherit from class
\code{"sparseMatrix"} as in package \code{Matrix}; not yet available for
\code{family="cox"})}

\item{y}{response variable. Quantitative for \code{family="gaussian"}, or
\code{family="poisson"} (non-negative counts). For \code{family="binomial"}
should be either a factor with two levels, or a two-column matrix of counts
or proportions (the second column is treated as the target class; for a
factor, the last level in alphabetical order is the target class). For
\code{family="multinomial"}, can be a \code{nc>=2} level factor, or a matrix
with \code{nc} columns of counts or proportions.  For either
\code{"binomial"} or \code{"multinomial"}, if \code{y} is presented as a
vector, it will be coerced into a factor. For \code{family="cox"}, \code{y}
should be a two-column matrix with columns named 'time' and 'status'. The
latter is a binary variable, with '1' indicating death, and '0' indicating
right censored. The function \code{Surv()} in package \pkg{survival}
produces such a matrix. For \code{family="mgaussian"}, \code{y} is a matrix
of quantitative responses.}

\item{...}{additional parameters passed to glmnet::glmnet.}
}
\description{
Not to be used directly by users
}
\keyword{internal}
