% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw.R
\name{IPCW}
\alias{IPCW}
\alias{brier_survival_weights}
\alias{roc_survival_weights}
\title{Inverse Probability of Censoring Weights (IPCW) for different metrics}
\usage{
brier_survival_weights(data, truth, analysis_time, model_fit)

roc_survival_weights(data, truth, analysis_time, model_fit)
}
\arguments{
\item{data}{A data frame with a column that corresponds to an object
created by \code{\link[survival:Surv]{survival::Surv()}} for right censored event times.}

\item{truth}{An unquoted column name for the \code{Surv} object.}

\item{analysis_time}{A vector of times at which the analysis will take
place. Note that, for the current censoring model that is used, these
values should be less than the largest observed event time.}

\item{model_fit}{A parsnip model object for a censored regression model. The
model should have been trained with parsnip version 1.0.4 or greater.}
}
\value{
Both functions return a column called \code{.censoring_weight}. Since the
Brier score weight calculations are dependent on the time of analysis, it
returns a list column with the weights for the ordered time values. The ROC
function is independent of analysis time so it returns a numeric column of
weights.

Additional classes are given to the  \code{.censoring_weight} column so that we
can check that the correct weighting scheme is used for the requested
metric.
}
\description{
These functions compute weights that are required for certain dynamic
measures of model performance when the outcome is a censored event time.
}
\details{
The weights use a "reverse Kaplan-Meier" curve to compute the estimates of
the probability that a data point is censored at a specific time. The
calculation assumes an uninformative censoring method. These values are
estimated with the training set data when the primary survival model is
trained.

Graf \emph{et al} (1999) use a combination of the observed event time and the time
at which the analysis is conducted. That paper is focused on the Brier
score metric. The function below encodes that logic.

Uno \emph{et al} (2007) show equations for estimating the area under the ROC curve
that only use the observed event times to compute the censoring probability.

Note that there are a few reasons that the weights may be estimated as \code{NA}.
First, is when the time of analysis is greater than the larges observed time in
the training set. Second, for the case of Graf \emph{et al} (1999), any data points
where the censored value is prior to the time of analysis (labeled as
"category 3") should not contribute to the analysis since their status at
the time of analysis is unknown.
}
\references{
Hajime Uno, Tianxi Cai, Lu Tian & L. J Wei (2007) Evaluating Prediction Rules
for t-Year Survivors With Censored Regression Models, \emph{Journal of the
American Statistical Association}, 102:478, 527-537. (see 5.1, equation 10)
}
