% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_hazards_data.R
\name{linear_pred_coxnet}
\alias{linear_pred_coxnet}
\title{A wrapper for predict() with coxnet models}
\usage{
linear_pred_coxnet(object, new_data, ...)
}
\arguments{
\item{object}{A fitted \verb{_coxnet} object.}

\item{new_data}{Data for prediction.}

\item{...}{Options to pass to \code{\link[glmnet:predict.glmnet]{glmnet::predict.coxnet()}}.}
}
\value{
A matrix.
}
\description{
A wrapper for predict() with coxnet models
}
\keyword{internal}
