% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_reg.R
\name{quantiles_flexsurvreg}
\alias{quantiles_flexsurvreg}
\title{Developer tools for flexsurv objects}
\usage{
quantiles_flexsurvreg(
  object,
  new_data,
  ...,
  quantile = (1:9)/10,
  interval = "none",
  level = 0.95
)
}
\arguments{
\item{object}{A model object.}

\item{new_data}{Data for prediction.}

\item{...}{Currently not used.}

\item{quantile}{Vector of quantiles.}

\item{interval}{Should the confidence interval be added? One of \code{"none"} or
\code{"confidence"}.}

\item{level}{Confidence level.}
}
\value{
A nested tibble.
}
\description{
Developer tools for flexsurv objects
}
\keyword{internal}
