% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_forest-ranger.R
\name{ranger_internal}
\alias{ranger_internal}
\alias{survival_time_ranger}
\alias{survival_prob_ranger}
\title{Internal helper functions for ranger survival models}
\usage{
survival_time_ranger(object, new_data)

survival_prob_ranger(object, new_data, eval_time)
}
\arguments{
\item{object}{A parsnip \code{model_fit} object resulting from
\link[parsnip:details_rand_forest_ranger]{rand_forest() with engine = "ranger"}.}

\item{new_data}{A data frame to be predicted.}

\item{eval_time}{A vector of times to predict the survival probability.}
}
\value{
\itemize{
\item \code{survival_time_ranger()}: A numeric vector of predicted survival times.
\item \code{survival_prob_ranger()}: A tibble with a list column of nested tibbles.
}
}
\description{
Internal helper functions for ranger survival models
}
\examples{
\dontshow{if (rlang::is_installed("ranger")) withAutoprint(\{ # examplesIf}
mod <- rand_forest() |>
  set_engine("ranger") |>
  set_mode("censored regression") |>
  fit(Surv(time, status) ~ age + ph.ecog, data = na.omit(lung))
survival_time_ranger(mod, new_data = lung[1:3, ])
\dontshow{\}) # examplesIf}
\dontshow{if (rlang::is_installed("ranger")) withAutoprint(\{ # examplesIf}
mod <- rand_forest() |>
  set_engine("ranger") |>
  set_mode("censored regression") |>
  fit(Surv(time, status) ~ age + ph.ecog, data = na.omit(lung))
survival_prob_ranger(mod, new_data = lung[1:3, ], eval_time = c(250, 100))
\dontshow{\}) # examplesIf}
}
\keyword{internal}
