% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_survival_prop.R
\name{survival_time_coxph}
\alias{survival_time_coxph}
\title{A wrapper for survival times with \code{coxph} models}
\usage{
survival_time_coxph(object, new_data)
}
\arguments{
\item{object}{A model from \code{coxph()}.}

\item{new_data}{Data for prediction}
}
\value{
A tibble
}
\description{
A wrapper for survival times with \code{coxph} models
}
\keyword{internal}
