% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_tree.R
\name{survival_time_survbagg}
\alias{survival_time_survbagg}
\title{A wrapper for survival times with \code{survbagg} models}
\usage{
survival_time_survbagg(object, new_data)
}
\arguments{
\item{object}{A model from \code{ipred::bagging()}.}

\item{new_data}{Data for prediction}
}
\value{
A tibble
}
\description{
A wrapper for survival times with \code{survbagg} models
}
\keyword{internal}
