% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model-survival.R
\name{survival_time_survfit_null}
\alias{survival_time_survfit_null}
\title{A wrapper for survival times with null models}
\usage{
survival_time_survfit_null(object, new_data)
}
\arguments{
\item{object}{A parsnip \code{model_fit} object resulting from a \code{null_model()}
with engine = "survival".}

\item{new_data}{Data for prediction.}
}
\value{
A numeric vector of predicted median survival times.
}
\description{
A wrapper for survival times with null models
}
\examples{
\dontshow{if (rlang::is_installed("survival")) withAutoprint(\{ # examplesIf}
mod <- parsnip::null_model() |>
  parsnip::set_engine("survival") |>
  parsnip::set_mode("censored regression") |>
  fit(survival::Surv(time, status) ~ ., data = survival::lung)
survival_time_survfit_null(mod, new_data = survival::lung[1:3, ])
\dontshow{\}) # examplesIf}
}
\keyword{internal}
