% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw.R
\name{unnest_survival_probs}
\alias{unnest_survival_probs}
\title{Tool for reorganizing dynamic survival predictions with censoring weights}
\usage{
unnest_survival_probs(data, ...)
}
\arguments{
\item{data}{A data frame with columns \code{.censoring_weight} and \code{.pred}}

\item{...}{Not currently used.}
}
\description{
Tool for reorganizing dynamic survival predictions with censoring weights
}
