% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_reg.R
\name{cox_reg}
\alias{cox_reg}
\alias{update.cox_reg}
\title{General Interface for Cox Models}
\usage{
cox_reg(mode = "censored regression")

\method{update}{cox_reg}(object, fresh = FALSE, ...)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible values for this model are "unknown", or "censored regression".}

\item{object}{A cox model specification.}

\item{fresh}{A logical for whether the arguments should be
modified in-place of or replaced wholesale.}

\item{...}{Not used for \code{update()}.}
}
\description{
\code{cox_reg()} is a way to generate a \emph{specification} of a model before
fitting and allows the model to be created using different packages in R.
}
\details{
For \code{cox_reg()}, the mode will always be "survival".

The model can be created using the \code{fit()} function using the following \emph{engines}:
\itemize{
\item \pkg{R}: \code{"survival"} (the default)
}
}
\section{Engine Details}{


Engines may have pre-set default arguments when executing the model fit call.
For this type of model, the template of the fit calls are:

\pkg{survival} engine

\preformatted{
survival::coxph(formula = missing_arg())
}
}

\examples{
library(survival)

cox_mod <-
  cox_reg() \%>\%
  set_engine("survival") \%>\%
  fit(Surv(time, status) ~ x, data = aml)
}
