% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_survival_prop.R
\name{coxnet_survival_prob}
\alias{coxnet_survival_prob}
\title{A wrapper for survival probabilities with coxnet models}
\usage{
coxnet_survival_prob(x, new_data, .times, training_data, output = "surv", ...)
}
\arguments{
\item{x}{A model from \code{glmnet()}.}

\item{new_data}{Data for prediction}

\item{.times}{A vector of integers for prediction times.}

\item{training_data}{A list of \code{x} and \code{y} containing the training data.}

\item{output}{One of "surv" or "haz".}

\item{...}{Options to pass to \code{\link[survival:survfit]{survival::survfit()}}}
}
\value{
A nested tibble
}
\description{
A wrapper for survival probabilities with coxnet models
}
\keyword{internal}
