% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_party.R
\name{details_decision_tree_party}
\alias{details_decision_tree_party}
\title{Decision trees via party}
\description{
\code{\link[party:ctree]{party::ctree()}} fits a conditional inference decision tree.
}
\details{
For this engine, there is a single mode: censored regression
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package (censored regression)}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{decision_tree() \%>\% 
  set_engine("party") \%>\% 
  set_mode("censored regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Decision Tree Model Specification (censored regression)
## 
## Computational engine: party 
## 
## Model fit template:
## censored::cond_inference_surv_ctree(formula = missing_arg(), 
##     data = missing_arg())
}

\code{\link[=cond_inference_surv_ctree]{cond_inference_surv_ctree()}} is a
wrapper around \code{\link[party:ctree]{party::ctree()}} (and other functions)
that makes it easier to run this model.
}

\subsection{Preprocessing requirements}{

This engine does not require any special encoding of the predictors.
Categorical predictors can be partitioned into groups of factor levels
(e.g. \verb{\{a, c\}} vs \verb{\{b, d\}}) when splitting at a node. Dummy variables
are not required for this model.
}

\subsection{References}{
\itemize{
\item Hothorn T, Hornik K, Zeileis A. 2006. Unbiased Recursive
Partitioning: A Conditional Inference Framework. \emph{Journal of
Computational and Graphical Statistics}, 15(3), 651–674.
\item Kuhn, M, and K Johnson. 2013. \emph{Applied Predictive Modeling}.
Springer.
}
}
}
\keyword{internal}
