% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_reg_flexsurv.R
\name{details_survival_reg_flexsurv}
\alias{details_survival_reg_flexsurv}
\title{Parametric survival regression}
\description{
\code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}} fits a parametric survival model.
}
\details{
For this engine, there is a single mode: censored regression
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{survival_reg() \%>\% 
  set_engine("flexsurv") \%>\% 
  set_mode("censored regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Parametric Survival Regression Model Specification (censored regression)
## 
## Computational engine: flexsurv 
## 
## Model fit template:
## flexsurv::flexsurvreg(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg())
}
}

\subsection{Other details}{

For this engine, stratification cannot be specified via
\code{\link[=strata]{strata()}}, please see the documentation of the
\code{\link{flexsurv}} package for alternative specifications.
}

\subsection{References}{
\itemize{
\item Jackson, C. 2016. \code{flexsurv}: A Platform for Parametric Survival
Modeling in R. \emph{Journal of Statistical Software}, 70(8), 1 - 33.
}
}
}
\keyword{internal}
