% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aorsf.R, R/partykit.R
\name{party_internal}
\alias{party_internal}
\alias{survival_prob_orsf}
\alias{survival_prob_cforest}
\alias{survival_prob_ctree}
\title{Internal helper functions for aorsf objects}
\usage{
survival_prob_orsf(object, new_data, time)

survival_prob_cforest(object, new_data, time, output = "surv")

survival_prob_ctree(object, new_data, time, output = "surv")
}
\arguments{
\item{object}{A model object from \code{partykit::ctree()} or \code{partykit::cforest()}.}

\item{new_data}{A data frame to be predicted.}

\item{time}{A vector of times to predict the survival probability.}

\item{output}{Type of output.}
}
\value{
A tibble with a list column of nested tibbles.

A tibble with a list column of nested tibbles.
}
\description{
Internal helper functions for aorsf objects

Internal helper functions for party objects
}
\examples{
library(aorsf)
lung_orsf <- na.omit(lung)
lung_orsf$status <- lung_orsf$status - 1 # aorsf is picky
aorsf <- orsf(Surv(time, status) ~ age + ph.ecog, data = lung_orsf, n_tree = 10)
survival_prob_orsf(aorsf, lung[1:3, ], time = 100)
library(partykit)
c_forest <- cforest(Surv(time, status) ~ age + ph.ecog, data = lung, ntree = 10)
survival_prob_cforest(c_forest, lung[1:3, ], time = 100)
c_tree <- ctree(Surv(time, status) ~ age + ph.ecog, data = lung)
survival_prob_ctree(c_tree, lung[1:3, ], time = 100)
}
\keyword{internal}
