% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{survival_prob_ctree}
\alias{survival_prob_ctree}
\alias{survival_prob_cforest}
\title{Internal helper functions for party objects}
\usage{
survival_prob_ctree(object, new_data, time)

survival_prob_cforest(object, new_data, time)
}
\arguments{
\item{object}{A model object}

\item{new_data}{A data frame to be predicted.}

\item{.time}{A vector of times to predict the survival probability.}
}
\value{
A tibble with a list column of nested tibbles.
}
\description{
Internal helper functions for party objects
}
\keyword{internal}
