% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_reg.R
\name{survival_prob_survreg}
\alias{survival_prob_survreg}
\alias{hazard_survreg}
\title{Internal function helps for parametric survival models}
\usage{
survival_prob_survreg(object, new_data, time)

hazard_survreg(object, new_data, time)
}
\arguments{
\item{object}{A \code{survreg} object.}

\item{new_data}{A data frame.}

\item{time}{A vector of time points.}
}
\value{
A nested tibble with column name \code{.pred}.
}
\description{
Internal function helps for parametric survival models
}
\keyword{internal}
