% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree.R
\name{survival_time_mboost}
\alias{survival_time_mboost}
\title{A wrapper for mean survival times with \code{mboost} models}
\usage{
survival_time_mboost(object, new_data)
}
\arguments{
\item{object}{A model from \code{blackboost()}.}

\item{new_data}{Data for prediction}
}
\value{
A tibble
}
\description{
A wrapper for mean survival times with \code{mboost} models
}
\keyword{internal}
