% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_tree-rpart.R
\name{survival_time_survbagg}
\alias{survival_time_survbagg}
\title{A wrapper for survival times with \code{survbagg} models}
\usage{
survival_time_survbagg(object, new_data)
}
\arguments{
\item{object}{A model from \code{ipred::bagging()}.}

\item{new_data}{Data for prediction}
}
\value{
A vector.
}
\description{
A wrapper for survival times with \code{survbagg} models
}
\examples{
library(ipred)
bagged_tree <- bagging(Surv(time, status) ~ age + ph.ecog, data = lung)
survival_time_survbagg(bagged_tree, lung[1:3, ])
}
\keyword{internal}
